Initial Commit
Hey Rowboat, this is called a commit. It's kind of like a snapshot of
your code -- You want to make them often, because it means if you fuck
up you can easily revert back.

When you're done with a "unit" of code -- a feature or something -- you
want to push it to the main repo (which is what I'm doing now). When
that happens, you'll merge with the existing repo. If there are merge
conflicts, you'll have to manually resolve them.

There are also things called branches, which I can tell you about
later. They're important, though. Ask me about it before you make any
major changes.

I'm using github for mac to give me a nice GUI for git, but you can do
it in the console as well. More on that later, if you decide to do this.